+++ .github/workflows/docker-build-and-publish.yml
@@ -5,7 +5,7 @@ on:
     branches: [master]
 
 jobs:
-  build-and-push:
+  build-and-push-amd64:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout repository
@@ -23,10 +23,36 @@ jobs:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
 
-      - name: Build and push Docker image
+      - name: Build and push Docker image for amd64
         env:
           IMAGE_NAME: dorowolf/akari-bot
           TAG: ${{ github.sha }}
         run: |
           docker buildx create --use
-          docker buildx build --platform linux/amd64,linux/arm64 -t $IMAGE_NAME:$TAG -t $IMAGE_NAME:latest --push .
+          docker buildx build --platform linux/amd64 -t $IMAGE_NAME:$TAG-amd64 -t $IMAGE_NAME:latest-amd64 --push .
+
+  build-and-push-arm64:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v2
+
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v2
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v2
+
+      - name: Login to DockerHub
+        uses: docker/login-action@v2
+        with:
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+
+      - name: Build and push Docker image for arm64
+        env:
+          IMAGE_NAME: dorowolf/akari-bot
+          TAG: ${{ github.sha }}
+        run: |
+          docker buildx create --use
+          docker buildx build --platform linux/arm64 -t $IMAGE_NAME:$TAG-arm64 -t $IMAGE_NAME:latest-arm64 --push .
