--- bots/lagrange/info.py
@@ -1,1 +1,0 @@
-client_name = 'Lagrange'
