+++ pyproject.toml
@@ -13,7 +13,7 @@ license = "AGPL-3.0"
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.12.0"
+python = "^3.9.0"
 Pillow = "^10.2.0"
 aiohttp = ">=3.9.3"
 beautifulsoup4 = "^4.11.1"
